+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/di/Koin.kt
@@ -7,8 +7,10 @@ import com.miam.kmmMiamCore.base.mvi.PointOfSaleStore
 import com.miam.kmmMiamCore.base.mvi.UserStore
 import com.miam.kmmMiamCore.component.itemSelector.ItemSelectorViewModel
 import com.miam.kmmMiamCore.component.preferencesSearch.PreferencesSearchViewModel
+import com.miam.kmmMiamCore.component.singletonFilter.SingletonFilterViewModel
 import com.miam.kmmMiamCore.handler.Basket.BasketHandler
 import com.miam.kmmMiamCore.handler.ContextHandler
+import com.miam.kmmMiamCore.handler.ToasterHandler
 import com.miam.kmmMiamCore.miam_core.data.datasource.MiamAPIDatasource
 import com.miam.kmmMiamCore.miam_core.data.repository.*
 import com.miam.kmmMiamCore.services.Analytics
@@ -57,6 +59,7 @@ val storeModule = module {
     single { PointOfSaleStore() }
     single { BasketHandler() }
     single { ContextHandler() }
+    single { ToasterHandler }
     single { ItemSelectorViewModel() }
     single { LikeStore() }
 }
@@ -64,4 +67,5 @@ val storeModule = module {
 val servicesModule = module {
     single { Analytics() }
     single { PreferencesSearchViewModel() }
+    single { SingletonFilterViewModel() }
 }